Fix: Correção completa dos componentes relacionados as regras

+++ src/renderer/src/components/Rule/index.tsx
@@ -1,72 +1,81 @@
-import { Button, Typography, useTheme } from "@mui/material";
+import { Typography, useTheme } from "@mui/material";
 import ContentWrapper from "../ContentWrapper";
-import { RuleProps } from "../../types/RulesProps";
 import { useSnackbar } from "../../context/SnackBarContext";
 import { useRuleStore } from "../../store/ruleStore";
 import { useRulePopupStore } from "../../store/popupRuleStore";
-const Rule = ({ id, name, extensions, description, isSystem, isActive }: RuleProps) => {
+import { FullRule } from "../../types/RuleWithDetails";
+import { useConfirmDialog } from "../../context/ConfirmDialogContext";
+import CrudButtons from "../CrudButtons";
+import { IConditionGroup } from "../../types/ConditionsType";
+
+const Rule = ({ id, name, description, isSystem, isActive, conditionsTree, action }: FullRule) => {
   const theme = useTheme();
   const { showMessage } = useSnackbar();
-  const { deleteRule, toggleActive } = useRuleStore();
+  const { deleteRule, toggleActive, duplicateRule } = useRuleStore();
   const { openPopup } = useRulePopupStore();
-  const ruleToEdit: RuleProps = {
+  const { showConfirm } = useConfirmDialog();
+  const ruleToEdit: FullRule = {
     id,
     name,
-    extensions,
     description,
     isActive,
+    isSystem,
+    conditionsTree,
+    action,
   };
-  const handleBtnClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    const { name, id } = event.currentTarget;
 
-    switch (name.toLowerCase()) {
-      case "edit":
-        openPopup("edit", ruleToEdit);
-        break;
-      case "duplicate":
-        handleDuplicate();
-        break;
-      case "delete":
-        handleDelete(parseInt(id));
-        break;
-      case "toggle":
-        handleToggle(parseInt(id));
-        break;
-      default:
-        break;
+  const handleDuplicate = async (ruleId: number) => {
+    const response = await duplicateRule(ruleId);
+    if (response.status) {
+      showMessage("Regra duplicada com sucesso", "success");
+    } else {
+      showMessage(response.message, "warning");
     }
   };
 
-  const handleDuplicate = () => {
-    console.log("Duplicar");
-  };
-
   const handleDelete = async (id: number) => {
-    await deleteRule(id);
-    showMessage("Regra excluída com sucesso.", "warning");
-  };
-
-  const handleToggle = async (id: number) => {
-    await toggleActive(id);
-    showMessage("Status da regra atualizado.", "info");
+    showConfirm(
+      {
+        title: "Excluir regra",
+        message: "Deseja realmente excluir esta regra? Essa ação não pode ser desfeita.",
+        confirmText: "Excluir",
+        cancelText: "Cancelar",
+      },
+      async () => {
+        await deleteRule(id);
+        showMessage("Regra excluída com sucesso.", "warning");
+      }
+    );
   };
 
   if (!id) return null;
 
+  const getExtensionsFromTree = (group: IConditionGroup): string[] => {
+    let extensions: string[] = [];
+    for (const child of group.children) {
+      if (child.type === "condition" && child.field === "fileExtension" && child.value) {
+        extensions.push(child.value);
+      } else if (child.type === "group") {
+        extensions = extensions.concat(getExtensionsFromTree(child));
+      }
+    }
+    return extensions;
+  };
+  const extensions = getExtensionsFromTree(conditionsTree).join(", ");
   return (
     <ContentWrapper
       title={name}
       titleTagType="h4"
       titleSize={22}
-      switchBtn={{ Action: () => toggleActive(id), value: isActive }}
       gap="1rem"
       bgColor="#2A3040"
+      switchBtn={{ Action: () => toggleActive(id), value: isActive }}
     >
       <Typography sx={{ color: theme.palette.text.secondary, fontSize: theme.typography.subtitle1 }}>
         {description}
       </Typography>
 
-      {extensions && (
+      {extensions.length > 0 && (
         <div style={{ display: "flex", gap: "1rem" }}>
           {extensions.split(",").map((item, index) => (
             <span
@@ -88,48 +97,12 @@ const Rule = ({ id, name, extensions, description, isSystem, isActive }: RulePro
 
       <div style={{ display: "flex", gap: 5 }}>
         {!isSystem && (
-          <>
-            <Button
-              id={id.toString()}
-              variant="outlined"
-              color="info"
-              sx={{ fontSize: 12, borderRadius: theme.shape.borderRadius }}
-              name="edit"
-              onClick={handleBtnClick}
-            >
-              Editar
-            </Button>
-
-            <Button
-              id={id.toString()}
-              variant="outlined"
-              color="warning"
-              sx={{
-                fontSize: 12,
-                borderRadius: theme.shape.borderRadius,
-                ":hover": { backgroundColor: "darkgoldenrod", color: "white" },
-              }}
-              name="duplicate"
-              onClick={handleBtnClick}
-            >
-              Duplicar
-            </Button>
-
-            <Button
-              id={id.toString()}
-              variant="outlined"
-              color="error"
-              sx={{
-                fontSize: 12,
-                borderRadius: theme.shape.borderRadius,
-                ":hover": { backgroundColor: "brown", color: "white" },
-              }}
-              onClick={handleBtnClick}
-              name="delete"
-            >
-              Excluir
-            </Button>
-          </>
+          <CrudButtons
+            id={id}
+            onEdit={() => openPopup("edit", ruleToEdit)}
+            onDelete={() => handleDelete(id)}
+            onDuplicate={() => handleDuplicate(id)}
+          />
         )}
       </div>
     </ContentWrapper>
